src/requirements.txt
asgiref==3.6.0
astroid==2.15.4
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.6
Django==4.1
django-extensions==3.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
drf-yasg==1.21.5
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.7.0
idna==2.10
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mypy==1.3.0
mypy-extensions==1.0.0
packaging==23.1
pkgutil-resolve-name==1.3.10
platformdirs==3.5.1
pluggy==1.0.0
psycopg2-binary==2.8.6
PyJWT==2.7.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyrsistent==0.19.3
pytest==7.3.1
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.25.1
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing-extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
wrapt==1.15.0
zipp==3.15.0